feat: register 폼과 register-optional 폼 분리
